Resolve comments, except for calculations regarding price
