--- examples/rust-sdk/lp-bot/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "lp-bot"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-clap = { version = "4.5.21", features = ["derive"] }
-colored = "2.0"
-orca_whirlpools = "0.4.0"
-orca_whirlpools_client = "0.4.3"
-orca_whirlpools_core = "0.4.3"
-serde_json = "1.0.133"
-solana-client = "1.18"
-solana-sdk = "1.18"
-spl-token-2022 = { version = "^3.0" }
-tokio = "1.41.1"
