+++ examples/rust-sdk/whirlpool_repositioning_bot/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@orca-so/whirlpools-example-rust-repositioning-bot",
+  "version": "0.0.1",
+  "scripts": {
+    "build": "cargo build",
+    "test": "cargo test --lib",
+    "format": "cargo clippy --fix --allow-dirty --allow-staged && cargo fmt",
+    "lint": "cargo clippy",
+    "clean": "cargo clean"
+  },
+  "devDependencies": {
+    "@orca-so/whirlpools-rust": "*",
+    "@orca-so/whirlpools-rust-client": "*",
+    "@orca-so/whirlpools-rust-core": "*"
+  }
+}
