+++ examples/rust-sdk/whirlpool_repositioning_bot/src/cli.rs
