+++ examples/ts-sdk/next/next-env.d.ts
