+++ examples/ts-sdk/next/next.config.js
@@ -12,7 +12,7 @@ const nextConfig = {
     config.plugins.push(
       new CopyWebpackPlugin({
         patterns: [{
-          from: "../../ts-sdk/core/dist/nodejs/orca_whirlpools_core_js_bindings_bg.wasm",
+          from: "../../../ts-sdk/core/dist/nodejs/orca_whirlpools_core_js_bindings_bg.wasm",
           to: "./server/app"
         }],
       })
