+++ package.json
@@ -26,7 +26,8 @@
     "rust-sdk/*",
     "ts-sdk/*",
     "legacy-sdk/*",
-    "examples/*",
+    "examples/ts-sdk/*",
+    "examples/rust-sdk/*",
     "docs/*"
   ],
   "lint-staged": {
